(RM4888) Abstract most supervisorctl commands
Push more calls to supervisorctl through a simple utils method. There's
still a fair amount of duplication, but this does allow us to pattern match
the output in a single place.

Leaving out `tail` at the moment because it is used interactively and
doesn't currently require any kind of output parsing.